Improve Github host entry on the Options page and make the injection handler event driven
