--- src/chrome/omnibox.js
@@ -1,2 +1,0 @@
-export const showDefaultSuggestion = description =>
-  chrome.omnibox.setDefaultSuggestion
