+++ src/chrome/permissions/github-host.js
@@ -1,0 +1,6 @@
+const githubHostPermissions = githubURL => ({
+  permissions: ["webRequest", "webRequestBlocking", "tabs"],
+  origins: [`*://${githubURL}/*`]
+});
+
+export default githubHostPermissions;
